Stop stringifying project worker shim
